pyproject.toml
[tool.poetry]
name = "fastapi-graphql-eventsourcing-async"
version = "0.1.0"
description = ""
authors = ["Chris <chris@uber5.com>"]
license = "UNLICENSED"
readme = "README.md"
packages = [{include = "fastapi_graphql_eventsourcing_async"}]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
fastapi = "^0.104.0"
strawberry-graphql = {extras = ["fastapi"], version = "^0.211.1"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
